requirements.txt
beautifulsoup4>=4.13.3
cloud-sql-python-connector>=1.18.0
langchain>=0.3.21
langchain-community>=0.3.20
langchain-core>=0.3.49
langchain-openai>=0.3.11
langchain-text-splitters>=0.3.7
pg8000>=1.31.2
pgvector>=0.4.0
python-dotenv>=1.1.0
requests>=2.32.3
SQLAlchemy>=2.0.40
google-cloud-secret-manager>=2.16.0
numpy==1.26.4
